Refactor legal info tooltip
Only show legal links if both label and url are configured
