+++ entry_types/scrolled/package/src/frontend/navigation/LegalInfoLink.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+export function LegalInfoLink(props) {
+  return (
+    <div>
+      {props.label && props.url &&
+      <a target="_blank" href={props.url}>
+        {props.label}
+      </a>
+      }
+    </div>
+  )
+}
