+++ entry_types/scrolled/package/src/frontend/navigation/LegalInfoMenu.js
@@ -1,18 +1,18 @@
 import React from 'react';
 import classNames from 'classnames';
 import headerStyles from "./AppHeader.module.css";
-import styles from "./ImprintAndPrivacy.module.css";
+import styles from "./LegalInfoMenu.module.css";
 import InfoIcon from "../assets/images/navigation/icons/information_icon.svg";
-import {ImprintAndPrivacyTooltip} from "./ImprintAndPrivacyTooltip";
+import {LegalInfoTooltip} from "./LegalInfoTooltip";
 
-export default function ImprintAndPrivacy(props) {
+export default function LegalInfoMenu(props) {
   return (
     <div>
       <a className={classNames(headerStyles.menuIcon, styles.infoIcon)}
-         data-tip data-for={'imprintAndPrivacyTooltip'} >
+         data-tip data-for={'legalInfoTooltip'} >
         <InfoIcon/>
       </a>
-      <ImprintAndPrivacyTooltip />
+      <LegalInfoTooltip />
     </div>
   )
 }
