+++ entry_types/scrolled/package/src/frontend/navigation/LegalInfoMenu.module.css
